Fix signed THP audio mixing

+++ src/THPSimple.cpp
@@ -877,7 +877,7 @@ void MixAudio(short* output, short* input, unsigned long samples)
                 volume = gTHPSimpleVolumeTable[static_cast<s32>(SimpleControl.unk_C4)];
 
                 mixedSample = static_cast<s32>(*input) +
-                              static_cast<s32>((static_cast<u32>(volume) * static_cast<s32>(*audioPtr)) >> 15);
+                              ((static_cast<s32>(volume) * static_cast<s32>(*audioPtr)) >> 15);
                 if (mixedSample < -0x8000) {
                     mixedSample = -0x8000;
                 }
@@ -887,7 +887,7 @@ void MixAudio(short* output, short* input, unsigned long samples)
                 *output = static_cast<s16>(mixedSample);
 
                 mixedSample = static_cast<s32>(input[1]) +
-                              static_cast<s32>((static_cast<u32>(volume) * static_cast<s32>(audioPtr[1])) >> 15);
+                              ((static_cast<s32>(volume) * static_cast<s32>(audioPtr[1])) >> 15);
                 if (mixedSample < -0x8000) {
                     mixedSample = -0x8000;
                 }
@@ -935,7 +935,7 @@ void MixAudio(short* output, short* input, unsigned long samples)
                 }
                 volume = gTHPSimpleVolumeTable[static_cast<s32>(SimpleControl.unk_C4)];
 
-                mixedSample = static_cast<s32>((static_cast<u32>(volume) * static_cast<s32>(*audioPtr)) >> 15);
+                mixedSample = (static_cast<s32>(volume) * static_cast<s32>(*audioPtr)) >> 15;
                 if (mixedSample < -0x8000) {
                     mixedSample = -0x8000;
                 }
@@ -944,7 +944,7 @@ void MixAudio(short* output, short* input, unsigned long samples)
                 }
                 *output = static_cast<s16>(mixedSample);
 
-                mixedSample = static_cast<s32>((static_cast<u32>(volume) * static_cast<s32>(audioPtr[1])) >> 15);
+                mixedSample = (static_cast<s32>(volume) * static_cast<s32>(audioPtr[1])) >> 15;
                 if (mixedSample < -0x8000) {
                     mixedSample = -0x8000;
                 }
